+++ src/tests/unit/twi/twi_slave_test.c
@@ -1,4 +1,4 @@
-#include "twi.h"
+#include "twi-slave.h"
 #include "gpio.h"
 
 int main(void) {
